refactor: use built-in map for in-memory storage
The hashicorp LRU cache implementation is not necessary here. The cache
will be cleared on every deployment anyway, it is not likely to be
long-running enough to need the expiration on various keys.

This means that this can be simplified to use the built-in map
implementation for Go and no longer require an extra dependency.