refactor: add routes as receiver funcs from server
Having these routes hang off from the `State` didn't really make sense.
They're for the HTTP server so should be one-level up, they still have
access to the `State` because this is part of the `PlaygroundServer`
already. Doing this removes the need of passing the `State` into every
function, as it can be accessed internally.